+++ src/ADCSense.cpp
@@ -26,25 +26,25 @@ void ADCSense::setup(adc_atten_t atten){
 	esp_err_t ret;
 	ret = esp_adc_cal_check_efuse(ESP_ADC_CAL_VAL_EFUSE_TP);
 	if (ret == ESP_ERR_NOT_SUPPORTED) {
-		Serial.printf( "Calibration scheme not supported, skip software calibration\n");
+		ESP_LOGE("ADCSense", "Calibration scheme not supported, skip software calibration");
 	} else if (ret == ESP_ERR_INVALID_VERSION) {
-		Serial.printf( "eFuse not burnt, skip software calibration\n");
+		ESP_LOGW("ADCSense", "eFuse not burnt, skip software calibration");
 	} else if (ret == ESP_OK) {
-		// Serial.printf( "eFuse burnt, software calibration can proceed\n");
+		// ESP_LOGI("ADCSense", "eFuse burnt, software calibration can proceed");
 
 		/*
 		esp_adc_cal_characteristics_t adc_chars;
 		esp_adc_cal_value_t val_type = esp_adc_cal_characterize(ADC_UNIT_1, ADC_ATTEN_DB_6, ADC_WIDTH_BIT_DEFAULT, ESP_ADC_CAL_VAL_DEFAULT_VREF, &adc_chars);
 		//Check type of calibration value used to characterize ADC
 		if (val_type == ESP_ADC_CAL_VAL_EFUSE_VREF) {
-			Serial.printf("eFuse Vref\n");
+			ESP_LOGI("ADCSense", "eFuse Vref");
 		} else if (val_type == ESP_ADC_CAL_VAL_EFUSE_TP) {
-			Serial.printf("Two Point\n");
+			ESP_LOGI("ADCSense", "Two Point");
 		} else {
-			Serial.printf("Default\n");
+			ESP_LOGI("ADCSense", "Default");
 		}
 		*/
-		Serial.printf("---- ADC: atten: %d channel: %d coeff: %1.3f\n", atten, ADCchannel, voltageDividerCoeff);
+		ESP_LOGI("ADCSense", "---- ADC: atten: %d channel: %d coeff: %1.3f", atten, ADCchannel, voltageDividerCoeff);
 
         adc1_config_channel_atten(ADCchannel, atten);
         esp_adc_cal_value_t val_type = esp_adc_cal_characterize(
@@ -65,7 +65,7 @@ void ADCSense::updatedADC(void){
 	raw = adc1_get_raw(ADCchannel);
 	// uint32_t raw_cal_mV = esp_adc_cal_raw_to_voltage(raw, adc_chars);
 	// mV = raw_cal_mV * voltageDividerCoeff;
-	// Serial.printf("VBAT sense reads %dmV --> %1.3fV VoltageDivider: %1.3fV - raw: %d\n", raw_cal_mV, (float)raw_cal_mV/1000, vBat/1000, raw);
+	// ESP_LOGI("ADCSense", "VBAT sense reads %dmV --> %1.3fV VoltageDivider: %1.3fV - raw: %d", raw_cal_mV, (float)raw_cal_mV/1000, vBat/1000, raw);
 
 
 	readTotal = readTotal - readings[readIndex];
@@ -78,7 +78,7 @@ void ADCSense::updatedADC(void){
 	raw_cal_mV = esp_adc_cal_raw_to_voltage(readTotal / ADC_SAMPLES, adc_chars);
 	mV = raw_cal_mV * (float)voltageDividerCoeff;
 	// float t = (float)voltageDividerCoeff;
-	// if (readIndex == 0) Serial.printf("VBAT sense reads %dmV --> %1.3fV VoltageDivider: %1.3fV --> %1.2fV- raw: %d voltageDividerCoeff: %1.4f\n", raw_mV, (float)raw_mV/1000, vBat/1000,vBat/1000, raw, t);
+	// if (readIndex == 0) ESP_LOGI("ADCSense", "VBAT sense reads %dmV --> %1.3fV VoltageDivider: %1.3fV --> %1.2fV- raw: %d voltageDividerCoeff: %1.4f", raw_mV, (float)raw_mV/1000, vBat/1000,vBat/1000, raw, t);
 
 
 }