+++ src/PowerManagement.h
@@ -13,6 +13,7 @@
 #include <FunctionalInterrupt.h>
 
 #include "ADCSense.h"
+#include <esp_log.h>
 #define VBUS_ADC_CHANNEL	ADC1_CHANNEL_0
 #define VBAT_ADC_CHANNEL	ADC1_CHANNEL_1
 
@@ -113,6 +114,7 @@ private:
 			if (chargingStatusCallbacks[i]) chargingStatusCallbacks[i]();
 		}
 		Serial.printf(" -- VBAT_STAT_SENSE_PIN changed to: %s --> ChargingStatus: %d\n",  vBatStat ? "true": "false", chargingStatus);
+		ESP_LOGI("PowerManagement", " -- VBAT_STAT_SENSE_PIN changed to: %s --> ChargingStatus: %d",  vBatStat ? "true": "false", chargingStatus);
 	}
 
 };